app/requirements.txt
Django==3.2.13
environs==7.3.1
djangorestframework
marshmallow
